+++ src/WebJobs.Script/Workers/Http/HttpWorkerProcess.cs
@@ -6,7 +6,6 @@ using System.Diagnostics;
 using Microsoft.Azure.WebJobs.Script.Diagnostics;
 using Microsoft.Azure.WebJobs.Script.Eventing;
 using Microsoft.Extensions.Logging;
-using Mono.Unix;
 
 namespace Microsoft.Azure.WebJobs.Script.Workers.Http
 {
@@ -19,7 +18,6 @@ namespace Microsoft.Azure.WebJobs.Script.Workers.Http
         private readonly string _scriptRootPath;
         private readonly string _workerId;
         private readonly WorkerProcessArguments _workerProcessArguments;
-        private readonly IEnvironment _environment;
 
         internal HttpWorkerProcess(string workerId,
                                        string rootScriptPath,
@@ -33,7 +31,7 @@ namespace Microsoft.Azure.WebJobs.Script.Workers.Http
                                        IMetricsLogger metricsLogger,
                                        IServiceProvider serviceProvider,
                                        ILoggerFactory loggerFactory)
-            : base(eventManager, processRegistry, workerProcessLogger, consoleLogSource, metricsLogger, serviceProvider, loggerFactory, httpWorkerOptions.Description.UseStdErrorStreamForErrorsOnly)
+            : base(eventManager, processRegistry, workerProcessLogger, consoleLogSource, metricsLogger, serviceProvider, loggerFactory, environment, httpWorkerOptions.Description.UseStdErrorStreamForErrorsOnly)
         {
             _processFactory = processFactory;
             _eventManager = eventManager;
@@ -42,7 +40,6 @@ namespace Microsoft.Azure.WebJobs.Script.Workers.Http
             _scriptRootPath = rootScriptPath;
             _httpWorkerOptions = httpWorkerOptions;
             _workerProcessArguments = _httpWorkerOptions.Arguments;
-            _environment = environment;
         }
 
         internal override Process CreateWorkerProcess()
@@ -60,31 +57,8 @@ namespace Microsoft.Azure.WebJobs.Script.Workers.Http
             workerContext.EnvironmentVariables.Add(HttpWorkerConstants.CustomHandlerPortEnvVarName, _httpWorkerOptions.Port.ToString());
             workerContext.EnvironmentVariables.Add(HttpWorkerConstants.CustomHandlerWorkerIdEnvVarName, _workerId);
             workerContext.EnvironmentVariables.Add(HttpWorkerConstants.FunctionAppRootVarName, _scriptRootPath);
-            Process workerProcess = _processFactory.CreateWorkerProcess(workerContext);
-            if (_environment.IsAnyLinuxConsumption())
-            {
-                AssignUserExecutePermissionsIfNotExists(workerProcess.StartInfo.FileName);
-            }
-            return workerProcess;
-        }
 
-        private void AssignUserExecutePermissionsIfNotExists(string filePath)
-        {
-            try
-            {
-                UnixFileInfo fileInfo = new UnixFileInfo(filePath);
-                if (!fileInfo.FileAccessPermissions.HasFlag(FileAccessPermissions.UserExecute))
-                {
-                    _workerProcessLogger.LogDebug("Assigning execute permissions to file: {filePath}", filePath);
-                    fileInfo.FileAccessPermissions |= FileAccessPermissions.UserExecute |
-                                                      FileAccessPermissions.GroupExecute |
-                                                      FileAccessPermissions.OtherExecute;
-                }
-            }
-            catch (Exception ex)
-            {
-                _workerProcessLogger.LogWarning(ex, "Error while assigning execute permission.");
-            }
+            return _processFactory.CreateWorkerProcess(workerContext);
         }
 
         internal override void HandleWorkerProcessExitError(WorkerProcessExitException httpWorkerProcessExitException)